requirements.txt
box2d-py==2.3.5
cffi==1.15.1
chess==1.7.0
click==8.1.7
cloudpickle==2.2.1
cmaes==0.10.0
cmake==3.27.0
contourpy==1.1.1
cycler==0.11.0
Farama-Notifications==0.0.4
filelock==3.12.4
fonttools==4.42.1
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.0
hanabi-learning-environment==0.0.4
inspyred==1.0.2
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.5
lit==16.0.6
llvmlite==0.40.1
magent2 @ file:///home/jawa17/Documents/Unitn/Master/ProjectCourse/Source/MAgent2
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
multi-agent-ale-py==0.1.11
networkx==3.1
numba==0.57.1
numpy==1.24.4
numpy-groupies==0.9.22
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.0
pettingzoo==1.22.4
Pillow==10.0.0
pycparser==2.21
pygame==2.3.0
pymunk==6.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
ribs==0.7.0
rlcard==1.0.5
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.13.0
shapely==2.0.1
six==1.16.0
sortedcontainers==2.4.0
sympy==1.12
termcolor==2.3.0
threadpoolctl==3.2.0
torch==2.0.1
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
